refactor(webapp): split DuoyinziStep into per-tab modules, dedupe fetch/search logic
DuoyinziStep.tsx had grown into a 1352-line god-component covering four
unrelated features (homograph list, GSUB rule browser/graph, IVS index,
phrase verification) with its own ad-hoc fetch helper and four copies of
debounced-search-plus-pagination state.

- Export request<T>() from api.ts and use it everywhere instead of a
  locally duplicated fetchJson/inline fetch (DuoyinziStep, VerifyTab)
- Extract useDebouncedPagedSearch, shared by GlyphsStep and the three
  duoyinzi tabs that had copy-pasted the same debounce/page-reset logic
- Split DuoyinziStep.tsx into pages/duoyinzi/{DuoyinziTab,GsubTab,IvsTab,
  VerifyTab,colors}.tsx; move their response types into types.ts
