+++ webapp/frontend/src/api.ts
@@ -9,7 +9,7 @@ import type {
   TaskState,
 } from './types'
 
-async function request<T>(url: string, init?: RequestInit): Promise<T> {
+export async function request<T>(url: string, init?: RequestInit): Promise<T> {
   const res = await fetch(url, {
     headers: { 'Content-Type': 'application/json' },
     ...init,
