+++ webapp/frontend/src/pages/duoyinzi/colors.ts
@@ -1,0 +1,8 @@
+// Reading-slot colors shared between the pattern tree (DuoyinziTab) and the
+// GSUB rule graph (GsubTab) so a given reading index reads the same in both views.
+export const SS_COLORS = [
+  'text-emerald-400 border-emerald-500/40 bg-emerald-500/10',
+  'text-sky-400 border-sky-500/40 bg-sky-500/10',
+  'text-amber-400 border-amber-500/40 bg-amber-500/10',
+  'text-rose-400 border-rose-500/40 bg-rose-500/10',
+]
